backend/requirements.txt
Django==2.2.2
django-cors-headers==2.4.0
djangorestframework==3.9.4
mysqlclient==1.3.13
pytz==2018.5